producer/infra/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.15.0
black==22.1.0
cattrs==1.10.0
click==8.0.4
constructs==10.0.82
isort==5.10.1
jsii==1.54.0
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
publication==0.0.3
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tomli==2.0.1
types-PyYAML==6.0.4
typing_extensions==4.1.1
